Fix click choices, refactor mensa bot
- Click cannot handle native Enums, so we have to go with namedtuples
  for now
- Mensa bot is now split into a class with an accompanying click script
